add in seperate component
