+++ frontend/src/components/Mint/MintItemContent.tsx
@@ -23,6 +23,7 @@ import {
 } from "@ionic/react";
 import { add, remove } from "ionicons/icons";
 import React, { Dispatch, SetStateAction, useState } from "react";
+import MintZoraLogic from "../MintZoraLogic";
 
 export interface MintItemContentProps {
   challenge: Challenge;
@@ -30,7 +31,15 @@ export interface MintItemContentProps {
 }
 
 const MintItemContent: React.FC<MintItemContentProps> = ({
-  challenge: { imageUrl, name, floorPrice, groupCount },
+  challenge: {
+    imageUrl,
+    name,
+    floorPrice,
+    groupCount,
+    mintingContractAddress,
+    tokenId,
+    chainId,
+  },
   setResult,
 }: MintItemContentProps) => {
   const ipfsImageUrl = convertIpfsImageUrl(imageUrl);
@@ -53,6 +62,8 @@ const MintItemContent: React.FC<MintItemContentProps> = ({
       setAmount(1);
     }
   };
+
+  console.log(mintingContractAddress, tokenId, chainId, "All of it");
 
   return (
     <IonGrid>
@@ -177,6 +188,13 @@ const MintItemContent: React.FC<MintItemContentProps> = ({
           >
             Mint {floorPrice} ETH
           </IonButton>
+          {chainId && tokenId && mintingContractAddress ? (
+            <MintZoraLogic
+              chainId={chainId}
+              tokenId={tokenId}
+              tokenContract={mintingContractAddress as `0x${string}`}
+            />
+          ) : null}
         </IonCol>
       </IonRow>
     </IonGrid>
